[modules-core][iOS] Introduce ExpoAppDelegateSubscriberManager
